mfd: cpcap-irq: Add debugfs file for CPCAP IRQ data
This change will add the file cpcap-irq to the debug fs directory. The file will
provide data for each of the cpcap interrupts. The data includes the interrupt
name, if the interrupt is enabled/disabled, if a handler has been registered
with the interrupt, and the number of times the interrupt has been generated.

Signed-off-by: Greg Meiste <w30289@motorola.com>