mdm6600: Add mdm6600 usb-serial driver.
This is a re-write of the Motorola mdm6600_modem driver.

The usb_serial_generic_* handlers actually cover the USB bulk to serial data
path. So that code disappears.

The USB interrupt data to TIOCMGET ioctl path does not appear to be used by
pppd in userspace, using a temporary BUG_ON() to confirm.

So we end up with very little but a hook to register generic usb-serial
handlers for the MDM6600 vendor id.

I have not spent time testing behavior in low-power-modes, it is likely I will
need to over-ride some of the generic handlers at that point.

Change-Id: I65fdd06c2764416f365995761aa653878c0e393b
Signed-off-by: Nick Pelly <npelly@google.com>