mdm6600: Enable the wake irq when ttyUSB4 is suspended
Instead of enabling it when the first opened port is suspended. It is
working because the modem only triggers this interrupt when all ttys are
suspended and the bus is suspended.

This fixes an issue when mdm6600_suspended_ports is not reset to 0 after
a disconnect/attach and the irq is never enabled.

Change-Id: I2a1a020ccd510b8c57c9f0b56988cb6de8aaff15
Signed-off-by: Benoit Goby <benoit@android.com>