mdm6600: Resume modem synchronously
If runtime pm auto-resumes the modem while the modem is being
disconnected, runtime pm might use the modem power spinlock after free.

Auto-resumes the modem synchronously and make sure there are no pending
wakeup requests before disconnect returns.

Don't request an auto-resume if the modem is suspended or this may
cause usb devices to be resumed in the wrong order.

Change-Id: I38bf4830d30bb8b7619e093e0ba1b3861d240aec
Signed-off-by: Benoit Goby <benoit@android.com>