mfd: cpcap: audio: Fix loss of audio over headset
When the 3.5mm headset is unplugged and plugged back in, cpcap gets into a
wierd state and audio is not routed to any speaker, neither headset nor
main speaker. This is a solution from Droid days. The cpcap sends a callback
to audio driver when such a plug/unplug is detected and forces the audio path
setup back to headset.

Change-Id: I91afda628e14ec20d589fc9b3e844d29f9f96fca
Signed-off-by: Praveen Bharathi <pbharathi@motorola.com>