power: bq24617: Ignore glitches on STAT pins
Due to the earthquake in Japan, the bq24617 is temporarily not available.
TI is providing another IC to mimic the bq24617, but unfortunately the
behavior is not identical. The STAT lines glitch due to a charger insertion.
The STAT line glitches obviously cause this driver to report the wrong
status to the system, causing the system to react incorrectly.

This patch will cause the STAT lines to be ignored for 500ms after a
charger insertion.

Change-Id: I465b36c2ab984ecda283fa8686bda87e6e09d5b5
Signed-off-by: Greg Meiste <w30289@motorola.com>