audio: tegra: Fix audio loss over headset
The audio callback function enforces cpcap reconfiguration when the
headset is unplugged and plugged back in less than 2 secs. In the short
interval between an unplug and plug, the registers are in the process of
unset and set and some of them end up getting programmed wrongly which
causes audio loss.

This patch ensures that all registers are wiped clean and reconfigured.

Signed-off-by: Praveen Bharathi <pbharathi@motorola.com>