radio_ctrl: wrigley: Fix powerup into flash mode
When the flash mode GPIO is asserted, the reset GPIO is not
connected, which is the only way to know the card powered on.
Currently, when trying to powerup into flash mode, the wrigley
driver returns failure, even though the operation succeeds. This
patch de-asserts the flash GPIO after giving the card time to see
it, therefore reconnecting the reset GPIO.

Change-Id: I65174f0e2b3e2b27462242ace0b1717223c03b9e
Signed-off-by: Greg Meiste <w30289@motorola.com>