usb: serial: mdm6600: Enable wake IRQ when modem is alive
Keep the IRQ disable when the irq fires, to prevent interrupt
storms when the modem resets or powers off.  But, enable it
after receiving data from the modem.

This [hack] is enabling an MDM6600 workaround, where it will
fire the wake IRQ 10 msec before each unsolicited tx.  This
prevents lost unsolicited messages, by ensuring the USB bus
will not be auto-suspending at the time the data is
pushed into the USB stack.

Change-Id: I3cc158772565d40bbf4117d29a7b7fae39a5633c
Signed-off-by: Chris Fries <C.Fries@motorola.com>