bridge: notifier called with the wrong device
If a new device is added to a bridge, the ethernet address of the
bridge network device may change. When the address changes, the
appropriate callback is called, but with the wrong device argument.
The address of the bridge device (ie br0) changes not the address
of the device being passed to add_if (ie eth0).

Change-Id: Idf106c889f34e06e2d47462dd883f30caad26eb8
Original-author: Stephen Hemminger <shemminger@vyatta.com>
Signed-off-by: Stephen Hemminger <shemminger@vyatta.com>