usb: modem: Don't request auto-resume in PREPARING state
If the modem is prepared, wait for the completion of the power
transition before deciding whether or not using autopm to wakeup the
modem.

Added a wakelock to qcusbnet to allow suspend only when the kthread is
idle.

Removed threshold time check from mdm6600 because the kernel now ignores
usb suspend errors since 2.6.39. This check should not be necessary if
autosuspend is correctly scheduled.

Bug: 5070702

Change-Id: I101fe8c31106318906120a9fa080467cdc28461a
Signed-off-by: Benoit Goby <benoit@android.com>