usb: core: Add usb_autopm_schedule_autosuspend
When USB resumes from system suspend, all the interfaces are resumed but
drivers can't queue an auto-suspend request because the interface is in
the RPM_SUSPENDED state. Add a function to the usb autopm API to queue a
request to be executed after the device is resumed. Drivers may delay
the request so that the device does not get suspended immediatly after
resume.

Bug: 5057045

Change-Id: I17ef98ed882646c9137d163cfdcad875c025344d
Signed-off-by: Benoit Goby <benoit@android.com>