[ARM] mfd: cpcap-audio: Fix headset bounce detection
CPCAP uses the headset detect interrupt as a mechanism
to prevent headset squeal and other artifacts.  It does
this by muting the headset amplifiers as soon as any
detection irq happens.

If the headset is quickly reinserted, the software must
turn the amplifiers back on.

If the headset is actually removed, we can rely on this
auto-mute functionality and let the software stack
decide where to route the audio next (rather than power
down the DAC/ADC.)

Change-Id: Ib145d3677c307175654cfd93b6e53ff1baf29e09