ARM: tegra: pm: keep a local cpu online mask for lp2 idle
Keep a local mask of cpus that are online, and mark a cpu
online before it is every brought out of reset.  Fixes a
problem when turning on the second cpu where the first cpu
may think that all cpus are idle while the second cpu is
still booting.  Locking around the online mask is
unnecessary because the the cpu that modifies the mask
must be outside of idle, so the online mask can never
equal the idle mask while it is being modified.

Change-Id: If703a16e5e7e3b0e316ab15083336fbe09814f36
Signed-off-by: Colin Cross <ccross@android.com>