v1.1.0 alpha changes for mach-tegra
-Experimental: now forging speedo id 4 and process id 2 so that EDP
limits are slightly raised and it narrows everyone down to a common DVFS
record for everyone. Raised Dynanic EDP governor to 67C (from 60C) to
give a little more room before edp is allowed to enable.

-Minor cpu voltage tweaks at the top end aiming at slightly better
battery for those that don't undervolt.

-Lowered the dvfs cold offsets from 50 to 25 for the top 4 slots.  This
will give folks a little more breathing room when undervolting and may
help cold performance a bit if your voltages are lowered close to their
lower limit.

-cpu transtition latency lowered - fairly certain that it only affects
OnDemand governor and not the default Interactive governor.

-Reverted most of the adjustments to the tegra 3 algorirthim for
bringing cpus online and offline. I think it livened it up, but at the
expense of battery.